Added style.css
CSS (Cascading Style Sheets) is a stylesheet language that enhances the presentation of HTML documents. It controls the layout, colors, and styles of elements on a webpage, allowing designers to create visually appealing and consistent user interfaces. CSS facilitates the separation of content from design, promoting efficient and flexible web development.